Small correction based on PR feedback

+++ include/g_local.h
@@ -1216,6 +1216,6 @@ qbool private_game_by_default(void);
 // set when match is starting, otherwise a normal spawn
 extern qbool initial_match_spawns;
 // highest number of players in all teams, used to allow (re)connect during match
-extern float maxPlayerCount;
+extern int maxPlayerCount;
 
 #define AUTOTRACK_POWERUPS_PREDICT_TIME 2
